[Node Telemetry] Move config to 'atlas-app-services' directory.
