+++ examples/node-telemetry/src/atlas-app-services/config.ts
